tools/requirements.txt
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
